refactor(heartbeat): extract wake contracts
