+++ src/infra/heartbeat-wake.ts
@@ -2,6 +2,14 @@
 import { runWithGatewayIndependentRootWorkAdmission } from "../process/gateway-work-admission.js";
 import { resolveTimerTimeoutMs } from "../shared/number-coercion.js";
 import { normalizeHeartbeatWakeReason } from "./heartbeat-reason.js";
+import type {
+  HeartbeatRunResult,
+  HeartbeatScheduledTask,
+  HeartbeatWakeHandler,
+  HeartbeatWakeIntent,
+  HeartbeatWakeOverride,
+  HeartbeatWakeSource,
+} from "./heartbeat-wake-contracts.js";
 import {
   abortHeartbeatWakeGeneration,
   type ActiveHeartbeatWakeTarget,
@@ -16,11 +24,14 @@ import {
 } from "./heartbeat-wake-target.js";
 
 export { getHeartbeatWakeAbortSignal } from "./heartbeat-wake-lifecycle.js";
-
-export type HeartbeatRunResult =
-  | { status: "ran"; durationMs: number }
-  | { status: "skipped"; reason: string; retryAtMs?: number }
-  | { status: "failed"; reason: string };
+export type {
+  HeartbeatRunResult,
+  HeartbeatScheduledTask,
+  HeartbeatWakeHandler,
+  HeartbeatWakeIntent,
+  HeartbeatWakeRequest,
+  HeartbeatWakeSource,
+} from "./heartbeat-wake-contracts.js";
 
 export const HEARTBEAT_SKIP_REQUESTS_IN_FLIGHT = "requests-in-flight";
 export const HEARTBEAT_SKIP_CRON_IN_PROGRESS = "cron-in-progress";
@@ -35,55 +46,6 @@ const RETRYABLE_GUARD_SKIP_REASONS = new Set(["not-due", "min-spacing", "flood"]
 export function isRetryableHeartbeatBusySkipReason(reason: string): boolean {
   return RETRYABLE_BUSY_SKIP_REASONS.has(reason);
 }
-
-export type HeartbeatWakeIntent = "scheduled" | "task" | "event" | "immediate" | "manual";
-
-export type HeartbeatWakeSource =
-  | "interval"
-  | "manual"
-  | "exec-event"
-  | "notifications-event"
-  | "cron"
-  | "hook"
-  | "background-task"
-  | "background-task-blocked"
-  | "acp-spawn"
-  | "session-state"
-  | "cli-watchdog"
-  | "restart-sentinel"
-  | "retry"
-  | "other";
-
-type HeartbeatWakeOverride = {
-  target?: string;
-  to?: string | undefined;
-  accountId?: string | undefined;
-};
-
-/** Cron-owned periodic work carried directly into a guarded heartbeat turn. */
-export type HeartbeatScheduledTask = {
-  jobId: string;
-  name: string;
-  prompt: string;
-};
-
-export type HeartbeatWakeRequest = {
-  source: HeartbeatWakeSource;
-  intent: HeartbeatWakeIntent;
-  reason?: string;
-  agentId?: string;
-  sessionKey?: string;
-  heartbeat?: HeartbeatWakeOverride;
-  /** Persisted cron monitor cadence carried with a scheduled heartbeat tick. */
-  scheduledEveryMs?: number;
-  /** Persisted cron monitor phase anchor carried with a scheduled heartbeat tick. */
-  scheduledAnchorMs?: number;
-  tasks?: readonly HeartbeatScheduledTask[];
-  /** Internal marker for work retained after a spacing/cooldown deferral. */
-  retainedWork?: boolean;
-};
-
-export type HeartbeatWakeHandler = (opts: HeartbeatWakeRequest) => Promise<HeartbeatRunResult>;
 
 let heartbeatsEnabled = true;
